LayoutTests/imported/w3c/web-platform-tests/html/capability-delegation/delegate-fullscreen-request-subframe-same-origin.https.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
Verifies that element.requestFullscreen() calls from a same-origin subframe without user activation work if and only if the top frame has user activation, regardless of whether it delegates the capability or not. https://wicg.github.io/capability-delegation/spec.html See wpt/html/user-activation/propagation*.html for frame tree user activation visibility tests.


PASS Fullscreen requests from a same-origin subframe fails without delegation from an opener with no user activation
PASS Fullscreen requests from a same-origin subframe succeeds without delegation from an opener with user activation
PASS Fullscreen requests from a same-origin subframe succeeds with delegation from an opener with user activation
